[refactor] splint applicatino module per product
