+++ src/app/component/error/error.module.ts
@@ -1,0 +1,15 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+// shared module
+import { SharedModule } from 'src/app/shared/shared.module';
+
+// components
+import { ErrorComponent } from './error.component';
+
+@NgModule({
+  declarations: [ErrorComponent],
+  imports: [CommonModule, SharedModule],
+  exports: [ErrorComponent],
+})
+export class ErrorModule {}
